+++ package.json
@@ -12,11 +12,12 @@
     "firebase": "^8.0.0",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
-    "react-redux": "^7.2.2",
     "react-hook-form": "^6.11.0",
+    "react-redux": "^7.2.2",
     "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.4",
     "redux": "^4.0.5",
+    "shortid": "^2.2.16",
     "uuid": "^8.3.1",
     "yarn": "^1.22.10"
   },
