+++ src/components/pages/DotDetail.jsx
@@ -5,6 +5,21 @@ import firebase from "../../firebase/firebase";
 export default function DotDetail() {
   const { id } = useParams();
   const [dot, set_dot] = useState();
+  // const dot = firebase
+  //   .firestore()
+  //   .collection("dots")
+  //   .doc(id)
+  //   .get()
+  //   .then((doc) => {
+  //     console.log(doc.data().title);
+  //     return doc.data();
+  //     // set_dot(doc.data());
+  //   });
+  // console.log(dot.handleFulfilled(value));
+
+  // .then((doc) => {
+  //   return doc;
+  // });
   useEffect(() => {
     firebase
       .firestore()
@@ -13,7 +28,7 @@ export default function DotDetail() {
         const dots = snapshot.docs.map((doc) => {
           return doc.data();
         });
-        set_dot(dots.find((dot) => dot.id == id));
+        set_dot(dots.find((dot) => dot.dotId == id));
       });
   }, []);
